Changed src layout
