+++ SConstruct
@@ -38,11 +38,11 @@ config = {
             '-Wno-unused-parameter',
             '-fdiagnostics-color',
             '-D__runtime_js__',
-            '-DRT_INC_ADDR_SPACE=\<kernel/x64/address-space-x64.h\>',
-            '-DRT_INC_CPU=\<kernel/x64/cpu-x64.h\>',
-            '-DRT_INC_IO=\<kernel/x64/io-x64.h\>',
-            '-DRT_INC_IRQ=\<kernel/x64/irqs-x64.h\>',
-            '-DRT_INC_PLATFORM=\<kernel/x64/platform-x64.h\>',
+            '-DRT_INC_ADDR_SPACE=\<kernel/arch/x64/address-space-x64.h\>',
+            '-DRT_INC_CPU=\<kernel/arch/x64/cpu-x64.h\>',
+            '-DRT_INC_IO=\<kernel/arch/x64/io-x64.h\>',
+            '-DRT_INC_IRQ=\<kernel/arch/x64/irqs-x64.h\>',
+            '-DRT_INC_PLATFORM=\<kernel/arch/x64/platform-x64.h\>',
             '-DRUNTIMEJS_PLATFORM_X64',
         ]),
         "release": set([
@@ -98,7 +98,17 @@ config = {
             'src',
             'src/arch',
             'src/kernel',
-            'src/kernel/x64',
+            'src/kernel/arch',
+            'src/kernel/arch/x64',
+            'src/kernel/boot',
+            'src/kernel/sys',
+            'src/kernel/sys/fs',
+            'src/kernel/sys/hw',
+            'src/kernel/sys/io',
+            'src/kernel/sys/memory',
+            'src/kernel/sys/thread',
+            'src/kernel/sys/v8',
+            'src/kernel/utils',
             'src/x64',
             'src/kernel/profiler',
             'src/common',
