+++ src/kernel/arch/platform.cc
