+++ src/kernel/arch/x64/acpi-x64.h
