+++ src/kernel/arch/x64/cpu-trampoline-x64.h
