+++ src/kernel/arch/x64/local-apic-x64.h
