+++ src/kernel/arch/x64/platform-x64.h
