+++ src/kernel/boot/kernel-main.cc
