+++ src/kernel/boot/multiboot.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "multiboot.h"
+#include <kernel/multiboot.h>
 #include <string>
 #include <kernel/boot-services.h>
