+++ src/kernel/boot/startup.cc
