+++ src/kernel/sys/hw/acpica-platform.cc
