+++ src/kernel/sys/memory/heap-snapshot.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "heap-snapshot.h"
+#include <kernel/heap-snapshot.h>
 #include <kernel/engines.h>
 
 namespace rt {
