+++ src/kernel/sys/thread/thread-manager.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "thread-manager.h"
+#include <kernel/thread-manager.h>
 #include <kernel/kernel.h>
 #include <kernel/engines.h>
 #include <kernel/platform.h>
