+++ src/kernel/sys/v8/native-thread.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "native-thread.h"
+#include <kernel/native-thread.h>
 
 namespace rt {
