+++ src/kernel/sys/v8/runtime-state.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "runtime-state.h"
+#include <kernel/runtime-state.h>
 
 namespace rt {
