+++ src/kernel/sys/v8/transport.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "transport.h"
+#include <kernel/transport.h>
 #include <kernel/template-cache.h>
 #include <kernel/object-wrapper.h>
 #include <kernel/thread.h>
